ml/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
onnx>=1.15.0
onnxruntime>=1.16.0
numpy>=1.24.0
scipy>=1.11.0
opencv-python>=4.8.0
ultralytics>=8.1.0
faiss-cpu>=1.7.4
tqdm>=4.65.0
